Fix navigation
The .active class was not being applied to the current page due to a
difference between the item.url and page.url. Which was due to the
trailing slash. Updating the permalink in the frontmatter to remove
the trailing slash resolved the issue but the permalink declaration is
also not necessary so I've removed it instead.